+++ src/pixelator/pna/analysis/segmentation/__init__.py
@@ -1,0 +1,8 @@
+"""Helpers for segmenting cell:cell conjugates.
+
+Copyright © 2026 Pixelgen Technologies AB.
+"""
+
+from pixelator.pna.analysis.segmentation.partition_counts import partition_counts
+
+__all__ = ["partition_counts"]